Load editors for votes in load_everything_for_edits
The code factored out into a load_everything_for_edits function in
commit 55fe59cff8af591ada077510a96482ea2f15836c appeared to load
editors for both votes and edit notes attached to an edit, but in
reality only did so for edit notes due to a syntactical bug. The only
place where it actually matters (displaying a single edit) had
slightly different code without the bug.

Change the function to always load editors for votes, too.