MBS-6887: Model coordinates with required fields
Coordinates for places are stored in the database as a point data
type, which has always both latitude and longitude; only the point as
a whole can be NULL. Therefore, points should be modelled with
mandatory latitude and longitude fields in the server code, too,
instead of indicating the absence of coordinates by setting latitude
and longitude to undefined.

This also makes it more straightforward to test for coordinates being
present or not, avoiding problems like MBS-7450 (zero coordinates
interpreted as no coordinates).