Don't try to avoid link_order conflicts
The implementation is causing more problems than it solves. And now
that you can see the exact link_order values in the UI, silently
changing what was entered can be very confusing. The code tried to make
a distinction between when the user did/did not intend for duplicate
positions, but it didn't work because of what information is all
included in the form submission.