MBS-7519: Remove z-index from PDF info tag
The tag marking thumbnails of PDF cover art had a z-index of 1000 with
the intention of layering it on top of the actual thumbnail. This made
it cover both the artwork viewer overlay and the pull-down menu,
however.

In fact, the z-index wasn't necessary at all, because positioned
elements (such as the "PDF" tag) are always layered on top of
non-positioned elements (such as the image thumbnail). It is therefore
removed.