MBS-7514: Reject user name "Deleted Editor #..."
A name of the form "Deleted Editor #123456" for a regular user is not
only confusing, but also prevents the account with ID 123456 from
being deleted, because the account renaming will fail. This applies to
other case variants, too, as the unique index on editor.name is case
insensitive. Therefore, this change makes the server reject
registration attempts for names of the type "Deleted Editor #...".