MBS-7354: Avoid ISE when a relationship has vanished
A relationship that was included in an entity edit form (mini
relationship editor) may have disappeared when the form is finally
submitted, namely if an edit to remove it has been applied or the edit
creating it has been cancelled in the meantime.

The controller did not anticipate this possibility and always expected
a relationship with known ID to exist. Hence, it would call a method
for an undefined relationship object, causing an internal server
error. This is avoided by exiting the loop iteration whenever a
(previously existing) relationship cannot be loaded.