Make type_id a required series field again
Since the field is disabled when there are items (and browsers don't
submit disabled fields), workarounds were added to fallback to whatever
the current series type is. Except there is no current series type for
*new* series, so the workarounds don't work. This makes the field
required again, and inserts a hidden input whose value is kept in sync
with the <select>, guaranteeing that a value is always submitted.