Allow removing dates from the mini rel. editor
Unfortunately, the mini relationship editor works from form data that
has not yet been validated (and normalized in the process). In
particular, the code in MB::S::Form::Field::PartialDate that sets
year/month/day for empty dates to undef is not run, nor is the code in
MB::S::Form::Field::DatePeriod to set the "ended" flag when an end
date is present.

Since refactoring the hook for the mini relationship editor is too
heavy a change for a post-schema-change planned hotfix, this commit
provides workarounds to allow removing dates. It fixes the tests from
the previous three commits.

This also fixes MBS-7545: Empty URL edits were created from the
release editor that differed in the normalized/non-normalized
representation of empty dates.