MBS-7556: Fix URL cleanup on edit URL pages
The URLCleanup function used to take a Knockout observable
"errorObservable" as a parameter and had a default to always treat it
as the empty string when the parameter was null. This has changed to
an errorCallback function parameter, but there was no default for it.
Therefore, the edit URL page that didn't provide a callback caused a
type error, and the cleanup code wasn't run.

This commit reinstates the previous behaviour: If no callback is given
when URLCleanup is called, it defaults to a function that always
returns the empty string.