Fix Artist::EditRelationships test, and ensure it still works in 5.18
Since perl 5.18 makes hashes inherently unsorted, it was possible for
attributes to end up reordered between the 'new' and 'old' data in an Edit
Relationship edit. This meant that when the test was updated, attributes were
included despite being identical when compared bagwise. This change ensures
that they're sorted before being compared for the sake of
Edit::WithDifferences, such that reordering alone will not result in an
entry.

(note that the 5.18 thing is a guess. but this is still correct)