MBS-7563/MBS-7567: Add sortname bubble iff there is a sortname
The initialize_guess_case function had a fixed list of entity types
having a sortname, and tried to initialize a sortname bubble (in
addition to the guess case bubble for the entity name) on all edit
forms for those entities. It did not discriminate between the main
entity edit form and the alias edit form.

This caused two problems:
- When sortnames were removed from labels and areas, they were not
  removed from the list. The function hence still tried to initialize
  a sortname bubble, causing a type error that prevented unrelated
  code from running. (MBS-7563)
- The aliases for entities without sortname (such as works and places)
  do have a sortname, but their edit form didn't have the sortname
  bubble initialized. (MBS-7567)

Both is fixed by dropping the fixed list, and instead testing if the
current form has a sortname field/bubble, to initialize them if and
only if they are present.