MBS-7595: Show release AC for "Remove disc ID" edits
In most cases (namely, when the medium still exists at the time of
display), the "Remove disc ID" edit presents release information from
the release that was loaded from the medium. It did not also load
artist credit information for the release, though, so this was not
available for display.

This is fixed by adding ArtistCredit as a post-load model for the
release. Additionally, the display data is changed so that it always
includes a medium entity (a placeholder one, if necessary) so that the
template can be simplified.