MBS-7615: Make sure URL edits are recognized as a change
Due to a scoping issue, the flag in the EditRelationships role
indicating whether any edits had been made to external links was
always false. If there were no edits to the core data or the other
relationships, the user would be redirected back to the edit form and
wrongly told that no changes had been made.