MBS-7303: Catch non-unique track IDs in medium edit
It has happened repeatedly that "edit medium" edits used the same
track ID several times in the new tracklist, causing all but the last
affected track to disappear when the edit is applied (MBS-7303).
Detect this situation when the edit is created and reject the invalid
tracklist. This might also help in finding the release editor bug that
is ultimately responsible.