Factor out date periods into an entity role
Most entities that support a date period (begin date/end date/ended)
did this via the Age role, although that didn't make sense for all of
them (Alias). The Link entity, on the other hand, had its own
implementation of a date period.

To remove the duplication, a new DatePeriod role is introduced, which
is used by Link, Alias, and Role::Age.