Get rid of the "msg" URL parameter
The editing interface for relationship types and relationship
attributes appended "?msg=created", "?msg=updated", "?msg=deleted" to
the URL that it redirected to after an edit. This was apparently done
(in commit 2565cde4f9ccf9c6899ed82cd56114b53c264e37) so that the new
page could show a banner with an appropriate message, but that feature
has never been implemented. It's also no longer necessary since a
message acknowledging the successful edit will be shown in any case,
so this commit removes the URL parameter altogether.