MBS-7739: Show full RG type in RG search results
When searching for release groups via indexed search, the full
(primary/secondary) types weren't displayed for the results. The
schema_fixup routine did contain code to interpret the primary and
secondary release group types, but it was only used for release groups
included in release search results, not for standalone release groups.