MBS-7771: Escape the content of textarea tags
Textarea fields that are prefilled with text containing HTML tags (or
otherwise using HTML reserved characters) hadn't this content escaped.
Most browsers accept this anyway and treat "<" as if it were "&lt;";
however, this means that it is treated the same as a literal "&lt;",
and both will be read as "<" when the form is submitted. This is a
problem in particular for link type documentation that may contain
HTML tags.

This is fixed by properly escaping all reserved characters, such that
"<" becomes "&lt;" and a literal "&lt;" becomes "&amp;lt;".