MBS-7753: Resolve macro parameter name conflict
Due to the way Template Toolkit handles variable scope, the use of a
"code" variable in the area/index.tt template caused the name of the
current user to be wrapped in a <code> tag. Rename the "code"
parameter of the _link_other_entity macro to "code_tag" to avoid this.