Add CSS class to relationships with pending edits
The display of relationships doesn't make it clear if any highlighting
for pending edits refers to the relationship itself or to its target.
As a first step towards a better display (MBS-7134), mark a
highlighting that refers to the relationship itself with the "mp-rel"
class in addition to "mp", so that the two cases can be distinguished
by user-side CSS or scripts.