MBS-7767: Prevent overriding a fixed entity type
In some cases, e.g. when editing an existing relationship, it is not
allowed to change the target entity type from within a relationship
editor dialogue. The type selection dropdown is accordingly disabled.
Still, it was possible to change the supposedly fixed entity type by
pasting the MBID of an entity with a different type. Check for this
case and reject the pasted MBID if the types don't match.