Move "view all aliases" link to the entity line
The edits affecting aliases (add/edit/remove) display a link to "view
all aliases" of the relevant entity, which used to be next to the name
of the alias being edited. However, it makes more sense to have it
next to the entity name instead. This also frees space on the alias
line for displaying additional information pertaining specifically to
the alias.