Fix ReferenceError on label bubble
In the templates, the way knockout works in letting you access variables
without referencing the owner is by using JavaScript's `with` statement,
so something like:

with ($data) { if (gid) { ... } }

But if gid isn't a property of $data, the JS will try to look on the window
object (or the above with statement if nested) and throw a ReferenceError
if it can't find a `gid` variable anyway. So the solution is just to use
`$data.gid` directly which will return `undefined` if it doesn't exist.

And the reason there's no gid is because a label isn't selected.