MBS-7790: Split "yes"/"no" strings for translation
As there are languages where "yes"/"no" cannot be translated the same
in connection with votes and with factual statements, add a context
hint "vote" where they are used in voting. Although "abstain" and
"approve" aren't used anywhere else, add the hint to them as well so
all strings can be treated the same.