Add limit clause for subscribed-entity edits queries
Normally, queries for edits (by an editor, for a certain entity, etc.)
have a limit of 500 (for the count of edits; the actual number of
edits to be displayed per page is lower). The queries for open edits
by a user's subscribed entities or editors did not, so this adds a
limit, to improve performance when there are more than 500 of them.