Move guess_search for entities to the template
The guess_search flag (indicating that the number of edits returned by
a query may be approximate only) was set from the controller role for
core entities' edit histories, whereas in all other cases it was set
from the template. Change this to always be set from the template.

This also fixes a bug where the list of edits for a collection was
missing the flag, because it uses the same template but is run from a
different controller.