MBS-7888: Keep forms on relationship editor page separate
Submitting any form from a relationship editor page started submission
preparation, even when it actually was a different form (search menu
in the page header). Limit this behaviour to the main edit form.

Also, the release relationship editor does this differently, and the
prepareSubmission function shouldn't ever be run for it.