MBS-7905: Fix JSON output for places without coordinates
Since commit 1bf6d79eef6e746faa6a17c41c5c1a6db49eebeb (MBS-6887), a
place's coordinates field may be undefined if there are no known
coordinates. The JSON serializer for ws/2 was not updated at the time,
so it would crash when applied to a place without coordinates.

This is fixed; at the same time, the output is changed to just have an
undefined value for the coordinates instead of a JSON object having
undefined latitude and longitude values.