Reorganize the string cleanup functions
The existing "trim" function for cleaning up a string is not
suitable for join phrases in artist credits because it also removes
any leading or trailing whitespace. Code that dealt with join phrases
therefore did its own cleanup which sometimes was equivalent to what
"trim" did and sometimes not.

This adds a "sanitize" function that does everything "trim" did,
except for removing leading and trailing whitespace, and implements
"trim" in terms of "sanitize" to ensure consistent behaviour. It also
uses "sanitize" where applicable.