Fix tests for Perl 5.20
Perl 5.20 includes the List::Util module in version 1.38 (as opposed
to version 1.27 in Perl 5.18), which in version 1.33 acquired a new
"all" function. Test::Deep by default re-exports this function if it
is available. Because many of our test modules have a test named
"all", a name clash occurs if they also import Test::Deep flatly.

This changes all uses of Test::Deep to only import the functions that
are actually needed, avoiding the clash.