MBS-7916: Fix "Set track durations" preview
The preview of the length changes on the "Set track durations" (from
CD TOC) page were broken for mediums having only a single track: If
the track had a length, it was still shown as having unknown length
("?:??"), and more tracks than the one existing appeared in the
comparison, with unknown lengths both for the medium and the TOC.

Both issues were caused by Template Toolkit interpreting the
single-element array of tracks returned by "all_tracks" as a scalar.
Consequently, the "size" virtual method operated on the track object
(a blessed hash) and reported the number of key-value pairs in the
hash, and trying to access the first element of the (non-)array
failed, returning an undefined value for the track length.