Un-inline code from the Unicode::Encoding plugin
Commit a88881a9002c430cac378c11ef1e1c47589716c5 copied some code from
the Unicode::Encoding catalyst plugin into the MBS codebase in order
to modify the behaviour in case decoding as UTF-8 fails, apparently
unaware that the handle_unicode_encoding_exception hook is provided
exactly for this purpose.

The plugin has now changed (rather, it has been incorporated into the
Catalyst core as of Catalyst::Runtime version 5.90040), and on systems
having a recent Catalyst runtime, the reference to the re-positioned
"$Catalyst::Plugin::Unicode::Encoding::CHECK" variable causes lots of
warning messages.

Remove the duplicated code and instead put the exception handler code
into the designated hook. This will work both with the plugin and the
in-core implementation.