Implement some core entity model methods via %ENTITIES
This adds a "_type" method to core entities' models, which returns the
entity type name, and adds a "mbid/multiple" flag to %ENTITIES for
entities that may have multiple MBIDs (via redirection). Together,
this allows for implementing the "_entity_class", "_table", and
"_gid_redirect_table" methods in Data::CoreEntity, in a generic form.
The individual implementations in Data::Area, Data::Artist etc. are
accordingly removed.