MBS-6011: Remove some direction marks from strings
Via users copying and pasting names, tracklists etc. from documents
containing mixed left-to-right and right-to-left text, strings with
unnecessary direction marks (left-to-right mark LRM, right-to-left
mark RLM) reach the database.

The direction marks only influence the rendering of adjacent
non-strong (weak or neutral) characters. When placed between two
strong characters or between a strong character and the boundary of
the string, they are therefore without effect and can be removed
during string sanitization.

This is implemented here, together with a removal of LRM/RLM from all
strings that don't contain any right-to-left characters. Analysis of
the strings with LRM or RLM currently in the database indicates that
those two cases cover most of the examples that occur in practice.