Test page for unused tags correctly
I have no idea what the previous test was doing since I can't find
where the strings "No artists," "No labels," etc. were coming from.

What it appears to be doing is inserting a "not-used" tag in the test
data, not using it, and then checking if the pages say something to
that effect, which I'm not sure makes sense since we remove unused
tags anyway.

This patch just makes the test check if the page for the not-used tag
now has a "Tag Not Used" heading.