MBS-7916: Fix "Set track durations" preview, again
The preview of the length changes on the "Set track durations" (from
CD TOC) page was broken for mediums having only a single track: If the
track had a length, it was still shown as having unknown length
("?:??"), and more tracks than the one existing appeared in the
comparison, with unknown lengths both for the medium and the TOC.

This had been fixed in commit 247ddbd629cdafe25d446e427405595a90780396
already, but the bug was reintroduced with the new pregap and data
tracks feature. It is caused by Template Toolkit's automatic type
conversion that can't be overridden, so the Perl code has to return an
array reference instead of a list to work around the TT design flaw.