Restrict the types of edits people can make with /ws/js/edit
This adds a whitelist of allowed edit types that /ws/js/edit can handle
at the moment, and additionally adds validation to ensure the endpoint
403s if a non-location-editor tries to edit an area, or a non-
relationship-editor tries to edit an instrument, in case we ever support
those edit types in the future.

To make restricting the relationship edits easier, a link type is now
required for *all* relationship edits in the request.