Refer to the nodejs-legacy package in INSTALL.md
Due to a name conflict, the main Node.js binary was renamed from
"node" to "nodejs" in Debian (https://bugs.debian.org/614907).
Packages to be installed via npm often don't expect this and fail to
install.

The long-term solution is to use system packages instead, though not
everything has been packaged yet; in particular, a Debian gulp package
is missing. So for now, recommend installing the nodejs-legacy package
which provides a symbolic link from "node" to "nodejs".