MBS-5959: Add unique indexes for subscriptions
Most of the tables storing subscription information
(editor_subscribe_artist etc.) had an index on (editor, entity) with
"_uniq" in its name, but it wasn't actually a unique index. Make these
indexes unique, as it doesn't make sense for an editor to be
subscribed to the same entity multiple times.