MBS-8064: Add missing link type documentation rows
When link types weren't added via the admin web interface, but through
migration scripts in recent schema changes, it was forgotten to add
the corresponding entries in documentation.link_type_documentation.
Trying to update these non-existing rows is a no-op; hence it is not
possible to change the documentation text from the web interface. This
affects some link types for places, instruments, series, and events.

The update script in this commit identifies all link types for which
there is no row in documentation.link_type_documentation and adds the
missing rows, with an empty documentation text.