Really set a database statement timeout
A statement timeout was supposed to be set for every database
session, but that never actually happened: It was conditional on the
connection being open already, which it never was because it is opened
on demand when the first query is run. Additionally, had the condition
ever been true, the do() would have failed because it requires
explicit transaction handling.