MBS-7984: Return to the edit search page after a timeout
If the edit search times out, catch this and show the edit search page
again with a message so that the user can try to narrow down the
search. This also requires reducing the database statement timeout
to less than the request timeout, so that the process still has time
to render the page.