MBS-8101: Don't immediately search for edits when refining
Add a query string parameter to the edit search page that tells it to
just display the search form pre-filled with the values from the
parameters, but not actually do the search yet. This is then used for
the "Refine this search" links so that users can first add further
conditions, instead of seeing the same edits as on the page they came
from.