MBS-8099: Don't focus text fields in edit search form
The edit search form gave newly-created text fields automatically
focus, to save the user a click or keystroke. However, many browsers
fire a "change" event on a dropdown whenever the user presses up/down,
even though the new content may only be an intermediate state: There
is no way to jump to a specific entry in the dropdown using the
keyboard. As a result, the focus was taken away from the type or
operator fields possibly before the user finished the selection,
forcing them to tab back multiple times. This rendered the form almost
unusable with a keyboard.

Because there appears to be no way to identify when the selection was
finished that would work reliably for mouse and for keyboard users,
and in all browsers, this change removes the programmatic focus change
completely. This also resolves MBS-4904, which asked not to give
focus to text fields on page load.