MBS-8097: Store ID/MBID in "Add relationship type" edits
"Add relationship type" edits didn't store the ID nor the MBID of the
newly created link type. Use Edit::Role::Insert to store them as
entity_id and entity_gid in the edit data, just like the edit types
for adding core entities do. This requires the insertion to be moved
from the "accept" method to the "insert" method because the former is
called too late, when the edit data is already written.