MBS-4928: Ignore approve votes in insert_votes_and_notes
It was possible for any editor (including non-autoeditors) to submit
approve votes for arbitrary edits. These would not cause the edit to
actually be approved, because that needs special logic only available
in its own method, but still confusingly recorded in the voting
history and voting statistics. Filter out these invalid votes in the
insert_votes_and_notes method.