Fix test data in anticipation of MBS-3998
Several tests for voting and adding edit notes used editors that
actually weren't entitled to vote or add edit notes. Give them a
registration date that is more than two weeks in the past and a number
of accepted edits so that they are no longer considered limited users.
The tests would otherwise have failed as soon as the restrictions for
limited users are enforced (MBS-3998).