Prevent users from finding edits for others' subscriptions
The edit search criterion "$entity is in my subscriptions" gets its
notion of whose subscriptions are "mine" from a hidden field in the
search form that is prefilled with the current user ID. However, a
user could easily change the field number to any other user ID and in
that way search for edits in that other user's subscriptions, even if
the other user had marked their subscriptions as private. This, in
turn, gave away which entities the other user was subscribed to.

As a provisional measure, check that the submitted user ID matches the
actual user and refuse to create the predicate otherwise.