MBS-6439: Give a hint about autocomplete paste-MBID feature
Instead of searching for an entity from an autocomplete field, it is
also possible to paste an MBID (or some longer string containing an
MBID, such as the URL for an entity) into the field. The existence of
this extremely useful feature isn't known widely enough, however. Add
a placeholder text to all autocomplete fields that touches on both
options.