MBS-8123: Fix jsonnew parsing of RG secondary types
Like for the other entity types, the new search server JSON output
format (jsonnew) uses a "secondary-types" array instead of
"secondary-type-list" for secondary types of release groups. The code
to parse this was not updated so that secondary types were missing
from search result pages for releases and release groups.