MBS-8122: Offer to change inline search method after failure
When an inline search failed, the user was given only the option to
retry the search, with the same method (indexed or direct search). As
the default is indexed search, in cases where the search server was
unavailable for more than a few seconds, this meant that inline search
was effectively unusable. Add a second option to retry the search with
the respective other method.