Retire tracklist_id for "Set track lengths" edits
The tracklist_id property of "Set track lengths" edits is no longer
used since shared tracklists were abolished. Even for older edits,
there is no way to translate the ID to the corresponding medium(s).
Hence, stop storing this in the edit data for new edits and handle it
via the "restore" mechanism for old edits.