MBS-2948: Show affected medium for "Set track lengths" edits
Edit display for "Set track lengths" edits only showed the affected
release(s), but not the specific medium if a release had multiple. For
newer edits that store a medium ID, load the medium, and use
medium_link to display something like "CD 3 on Some Release". (For
older edits, the information is not available, so this can't be
done.)