MBS-7900: Prevent duplicate tags in ws/2 responses
In ws/2 release lookups, the "tags" include parameter causes tags to
be loaded for artists in artist credits both on the release and on
recordings, if these are included as well. The loading happens in
separate steps. If an artist appears both in the release AC and in a
recording AC and their tags are hence loaded twice, the tags from the
second round were appended to the existing list of tags for that
artist. This caused all tags for such artists to appear twice in the
final output.

To fix this, when loading tags, replace any existing list of tags for
an artist instead of appending to it.