MBS-3452: Consistently order annotations on merge
When entities with annotations were merged, the order of the
annotations in the concatenation was essentially random, as they were
read from a hash. This change makes the order deterministic (the same
in which the entity IDs were passed to the merge method), and in
particular lets the annotation from the merge target come first.