MBS-8095: Process subscription emails in batches
The script for sending out notifications about edits to editors'
subscribed entities used to load the data for all editors with
subscriptions into memory, contributing to the script being aborted in
low-memory situations. Instead, load the data for only a limited
number of editors at a time, process and discard it before loading the
next batch.