MBS-8095: Don't cache edits for obscure entities
The script for sending out notifications about edits to editors'
subscribed entities used to cache the edits for each entity in RAM.
This contributed to an increasing memory consumption over the script's
running time, finally causing it to be aborted for lack of memory.

Since the majority of entities (currently over 90%) only has one
subscriber, caching all this data is of limited use. This makes it so
that after processing each batch of 1000 editors, all entities that
haven't been seen at least twice are removed. Due to the effect known
as birthday paradox, this still ensures that most of the common
entities will eventually remain in the cache, though not necessarily
the first time they appear. (Greater than 50% probability for an
entity subscribed to by 11 editors; greater than 90% for 19.) On the
other hand, the long tail of single-subscriber entities is removed
from the cache with certainty.