MBS-8145: Treat database timeouts differently from ISEs
Database timeouts are unfortunately not a rare occurrence for certain
types of requests. They used to be treated as an internal server
error, i.e. an email was sent to MetaBrainz employees about each
timeout, and a rather technical error page with a stack trace shown to
the user.

Instead, suppress error emails for database timeouts (but log them)
and show a dedicated error page suggesting to reload the page, with a
503 status code (service unavailable). The error page is targeted at
non-technical users, but has in a hidden element the same debugging
information that is on an ISE page.