Make /relationship-editor attribute code less fragile
This firstly fixes a bug where you couldn't remove all attributes from a
relationship using the /relationship-editor endpoint. This was because the
`attributes` field worked by replacing the attributes on a relationship
entirely with the newly-submitted set, and there was no way to submit an empty
set. (Submitting nothing would leave the attributes alone.)

This introduces a breaking change by requiring all removals to be explicit. A
`removed` field must be sent with any attributes that are to be removed. Any
other attributes submitted are merged into the existing ones. So they'll either
be added, ignored, or updated if a different credited_as/text_value is sent.

Since /relationship-editor works by forwarding its data to /ws/js/edit, it made
sense to bring these changes over to there for consistency.