"Remove medium" edit: Load medium for display, if available
For the "Remove medium" edit display, try to load the affected medium
by its ID. This has the advantage that, as long as the medium exists
(because the edit is still open or has failed), its current position,
title etc. will be displayed instead of the state frozen when the edit
was created. This also permits use of the existing medium_link macro.