Remove index check on previousTrackAtThisPosition
I added this because I was thinking that moving the track should invalidate
`previousTrackAtThisPosition`, since, well, it's in a different position after
moving it. But it's correct to keep `previousTrackAtThisPosition`, (though the
name might be misleading) since moving a track should be equivalent to moving
the previous track, too, assuming they're actually the same, which is the
purpose of tracking this data.