Fix a warning in Data::Work::find_artists
The "map" call produced a list of hash references, one for each ID,
instead of a list of ID/inner hash reference pairs as was obviously
intended. This caused a warning about the hash initialization list
having an odd number of elements if there was an odd number of IDs.

The resulting data isn't used later at all, so simply remove the code.