MBS-6763: Merge "ended" flag
The "ended" flag was not correctly merged as a boolean attribute;
instead, the value of the merge target was kept (unless an end date
forced the flag to be set). Add a merge_date_period helper function
that merges begin and end dates and the ended flag correctly, and use
it for all entities that have date periods.