MBS-8162: Use relationship seeding for "Add work" link
The "Add work" link in the sidebar on artist pages had an "artist" URL
parameter containing the MBID of the artist, but this parameter was
subsequently ignored. Use relationship seeding instead to add a
relationship to the artist, for which the user can then choose an
appropriate type (e.g. "composed").