MBS-8300: Display medium title "0" in release view
The position_and_name method used for generating the "CD 1: Title"
text in the release view left out the title when it was false,
including when it was "0". Use a new medium_description macro instead,
split out from the medium_link macro which already worked correctly.
This also removes some duplication in code and translation strings.