MBS-8266: Make medium titles VARCHAR NOT NULL
Medium titles (medium.name) were limited to a length of 255 characters
and could be NULL. This was inconsistent with other entity names,
which don't have a length limit, and also with other cases of optional
string values, where the empty string is used to indicate the absence
of a value (e.g., disambiguation comment on all types of entities,
address for places); there is also no semantic distinction that could
be expressed by a NULL value as opposed to an empty string.

The type of medium.name is therefore changed to VARCHAR NOT NULL, and
existing NULL values are updated to the empty string. This implicates
changes to the edit data for some medium edits, disallowing undefined
or missing name values, and to the corresponding JavaScript models.