MBS-8330: Undo HTML encoding before comparing names
For marking name variations, we compare the display name with the name
of the entity. When the display name is from a diff and already in
HTML, it is necessary not only to remove any <span> tags before the
comparison but also to undo the conversion of HTML reserved characters
into entities. The latter wasn't done before this commit; therefore,
names containing one of the reserved characters were treated as a
variation even when they were, in fact, identical to the entity name.