MBS-8321: Readding same information in the relationship editor gives unexpected results
There were two problems: one was that, in the edit data, 'ended' wasn't being
set to true if there was an end date. This didn't matter for the server side
of things since it gets set anyway, but it throws off the isDuplicate
comparison locally.

The second problem was that mergeDates didn't properly handle comparisons
between numbers and strings.