Change statistics query counting the MBIDs
Change the statistics query that counts all the MBIDs in use, so that
it counts each entity type separately and adds the numbers instead of
producing a (virtual) list of all MBIDs and counting those. The former
appears to require less resources (though the reason for that is not
known).